+++ server/src/main/java/com/example/echo/location/dto/GeocodingResult.java
@@ -2,14 +2,15 @@ package com.example.echo.location.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Getter;
+import lombok.Data;
 import lombok.NoArgsConstructor;
 
-@Getter
+@Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class GeocodingResult {
+
     private String placeName;
     private String address;
 }
